3863: Fixes to queue handling
