+++ .env
@@ -101,3 +101,9 @@ INVOICE_ENTRY_ACCOUNTS='{
 ###< symfony/mailer ###
 
 SYNC_MONITORING_URL=
+
+###> symfony/lock ###
+# Choose one of the stores below
+# postgresql+advisory://db_user:db_password@localhost/db_name
+LOCK_DSN=flock
+###< symfony/lock ###
