+++ config/packages/lock.yaml
@@ -1,0 +1,2 @@
+framework:
+    lock: '%env(LOCK_DSN)%'
