+++ docker-compose.server.override.yml
@@ -9,7 +9,7 @@ services:
         stop_grace_period: 20s
         environment:
             - APP_SUPERVISOR_COMMAND=/app/bin/console messenger:consume --env=prod --no-debug --time-limit=900 --failure-limit=1 async scheduler_default
-            - APP_SUPERVISOR_WORKERS=4
+            - APP_SUPERVISOR_WORKERS=1
             - APP_SUPERVISOR_USER=deploy
             - PHP_MAX_EXECUTION_TIME=60
             - PHP_MEMORY_LIMIT=256M
