+++ src/Command/QueueSyncCommand.php
@@ -25,9 +25,6 @@ class QueueSyncCommand extends Command
     public function __construct(
         private readonly SynchronizationJobRepository $synchronizationJobRepository,
         private readonly MessageBusInterface $bus,
-        private readonly HttpClientInterface $client,
-        private readonly LoggerInterface $logger,
-        private readonly string $monitoringUrl,
     ) {
         parent::__construct($this->getName());
     }
@@ -38,15 +35,6 @@ class QueueSyncCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        // Call monitoring url if defined.
-        if ('' !== $this->monitoringUrl) {
-            try {
-                $this->client->request('GET', $this->monitoringUrl);
-            } catch (\Throwable $e) {
-                $this->logger->error('Error calling monitoringUrl: '.$e->getMessage());
-            }
-        }
-
         $job = new SynchronizationJob();
         $job->setStatus(SynchronizationStatusEnum::NOT_STARTED);
         $this->synchronizationJobRepository->save($job, true);
