+++ src/MessageHandler/SynchronizeHandler.php
@@ -9,8 +9,11 @@ use App\Repository\DataProviderRepository;
 use App\Repository\ProjectRepository;
 use App\Repository\SynchronizationJobRepository;
 use App\Service\DataSynchronizationService;
+use Psr\Log\LoggerInterface;
+use Symfony\Component\Lock\LockFactory;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
 use Symfony\Component\Messenger\Exception\UnrecoverableMessageHandlingException;
+use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 #[AsMessageHandler]
 class SynchronizeHandler
@@ -20,12 +23,26 @@ class SynchronizeHandler
         private readonly DataProviderRepository $dataProviderRepository,
         private readonly ProjectRepository $projectRepository,
         private readonly SynchronizationJobRepository $synchronizationJobRepository,
+        private readonly HttpClientInterface $client,
+        private readonly LoggerInterface $logger,
+        private readonly string $monitoringUrl,
+        private readonly LockFactory $factory,
     ) {
     }
 
     public function __invoke(SynchronizeMessage $message): void
     {
+        // TTL for the lock is two hours.
+        $lock = $this->factory->createLock('sync-all', ttl: 60 * 60 * 2);
+
         try {
+            if (!$lock->acquire()) {
+                $this->logger->warning("Lock could not be acquired.");
+                return;
+            }
+
+            $this->logger->info('sync-all lock acquired');
+
             $job = $this->synchronizationJobRepository->find($message->getSynchronizationJobId());
 
             if (null === $job) {
@@ -106,6 +123,15 @@ class SynchronizeHandler
             $job->setStatus(SynchronizationStatusEnum::DONE);
             $job->setEnded(new \DateTime());
             $this->synchronizationJobRepository->save($job, true);
+
+            // Call monitoring url if defined.
+            if ('' !== $this->monitoringUrl) {
+                try {
+                    $this->client->request('GET', $this->monitoringUrl);
+                } catch (\Throwable $e) {
+                    $this->logger->error('Error calling monitoringUrl: '.$e->getMessage());
+                }
+            }
         } catch (\Throwable $throwable) {
             $job = $this->synchronizationJobRepository->find($message->getSynchronizationJobId());
 
@@ -117,6 +143,9 @@ class SynchronizeHandler
 
             // Avoid retrying job.
             throw new UnrecoverableMessageHandlingException($throwable->getMessage(), (int) $throwable->getCode() ?? 0);
+        } finally {
+            $lock->release();
+            $this->logger->info('sync-all lock released.');
         }
     }
 }
